+++ crates/ruma-signatures/tests/tests.rs
@@ -8,8 +8,10 @@ static PKCS8_ED25519_DER: &[u8] = include_bytes!("./keys/ed25519.der");
 fn add_key_to_map(public_key_map: &mut PublicKeyMap, name: &str, pair: &Ed25519KeyPair) {
     let sender_key_map = public_key_map.entry(name.to_owned()).or_default();
     let encoded_public_key = Base64::new(pair.public_key().to_vec());
-    let version =
-        ServerSigningKeyId::from_parts(SigningKeyAlgorithm::Ed25519, pair.version().into());
+    let version = ServerSigningKeyId::from_parts(
+        SigningKeyAlgorithm::Ed25519,
+        pair.version().try_into().unwrap(),
+    );
 
     sender_key_map.insert(version.to_string(), encoded_public_key);
 }